Repackage Spanner coders
